fill the circle, fill the rhyme

+++ src/shapes.h
@@ -62,7 +62,7 @@
     void SR_DrawCirc(
         SR_Canvas *canvas,
         SR_RGBAPixel colour,
-        unsigned short x,
-        unsigned short y,
-        unsigned short r);
+        int x,
+        int y,
+        unsigned long r);
 #endif
